+++ applications/fishing-map/src/features/dataviews/dataviews.mock.ts
@@ -1,4 +1,4 @@
-import { Dataview } from '@globalfishingwatch/api-types/dist'
+import type { Dataview } from '@globalfishingwatch/api-types/dist'
 
 export const dataviews: Dataview[] = []