+++ applications/vessel-history/src/features/dataviews/dataviews.utils.ts
@@ -30,12 +30,20 @@ export const getVesselDataviewInstanceId = (vesselId: string) => `${VESSEL_LAYER
 
 export const getVesselDataviewInstance = (
   vessel: { id: string },
-  { trackDatasetId, infoDatasetId, eventsDatasetsId }: VesselInstanceDatasets
+  { trackDatasetId, infoDatasetId, eventsDatasetsId }: VesselInstanceDatasets,
+  akaVessels: { id: string }[] = []
 ): DataviewInstance<Generators.Type> => {
+  // Build list of unique vessel ids to merge
+  // sorted alphabetically so that regardless of the order
+  // in which the user selected the vessels
+  // the request url is the same
+  const akaVesselsIds = Array.from(
+    new Set([vessel.id].concat(akaVessels.map((v) => v.id)).sort((a, b) => (a > b ? 1 : -1)))
+  ).join(',')
   const datasetsConfig: DataviewDatasetConfig[] = [
     {
       datasetId: trackDatasetId,
-      params: [{ id: 'vesselId', value: vessel.id }],
+      params: [{ id: 'vesselId', value: akaVesselsIds }],
       endpoint: EndpointId.Tracks,
     },
     {
@@ -48,7 +56,7 @@ export const getVesselDataviewInstance = (
     eventsDatasetsId.forEach((eventDatasetId) => {
       datasetsConfig.push({
         datasetId: eventDatasetId,
-        query: [{ id: 'vessels', value: vessel.id }],
+        query: [{ id: 'vessels', value: akaVesselsIds }],
         params: [],
         endpoint: EndpointId.Events,
       })