+++ applications/vessel-history/src/features/profile/components/Info.tsx
@@ -1,9 +1,10 @@
 import React, { Fragment, useEffect, useMemo, useState } from 'react'
 import cx from 'classnames'
+import { event as uaEvent } from 'react-ga'
 import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import ImageGallery from 'react-image-gallery'
-import { DateTime } from 'luxon'
+import { DateTime, Interval } from 'luxon'
 import { Authorization } from '@globalfishingwatch/api-types'
 import { Button, IconButton } from '@globalfishingwatch/ui-components'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
@@ -15,6 +16,7 @@ import { OfflineVessel } from 'types/vessel'
 import {
   selectDataset,
   selectTmtId,
+  selectUrlAkaVesselQuery,
   selectVesselId,
   selectVesselProfileId,
 } from 'routes/routes.selectors'
@@ -28,6 +30,7 @@ interface InfoProps {
   vessel: VesselWithHistory | null
   lastPosition: any
   lastPortVisit: any
+  onMoveToMap: () => void
 }
 
 const Info: React.FC<InfoProps> = (props): React.ReactElement => {
@@ -40,22 +43,44 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
   const vesselTmtId = useSelector(selectTmtId)
   const vesselDataset = useSelector(selectDataset)
   const vesselProfileId = useSelector(selectVesselProfileId)
+  const akaVesselProfileIds = useSelector(selectUrlAkaVesselQuery)
   const offlineVessel = useSelector(selectCurrentOfflineVessel)
   const { dispatchCreateOfflineVessel, dispatchDeleteOfflineVessel, dispatchFetchOfflineVessel } =
     useOfflineVesselsAPI()
+  const isMergedVesselsView = useMemo(
+    () => akaVesselProfileIds && akaVesselProfileIds.length > 0,
+    [akaVesselProfileIds]
+  )
 
   useEffect(() => {
     dispatchFetchOfflineVessel(vesselProfileId)
   }, [vesselProfileId, dispatchFetchOfflineVessel])
 
   const onDeleteClick = async (data: OfflineVessel) => {
+    const now = DateTime.now()
+    const savedOn = DateTime.fromISO(data.savedOn);
+    const i = Interval.fromDateTimes(savedOn, now);
+    uaEvent({
+      category: 'Offline Access',
+      action: 'Remove saved vessel for offline view',
+      label: JSON.stringify({ page: 'vessel detail' }),
+      value: Math.floor(i.length('days'))
+    })
     setLoading(true)
-    await dispatchDeleteOfflineVessel(data.profileId)
+    await dispatchDeleteOfflineVessel(data)
     setLoading(false)
   }
 
   const onSaveClick = async (data: VesselWithHistory) => {
     setLoading(true)
+    uaEvent({
+      category: 'Offline Access',
+      action: 'Save vessel for offline view',
+      label: JSON.stringify({
+        gfw: vesselId,
+        tmt: vesselTmtId
+      })
+    })
     await dispatchCreateOfflineVessel({
       vessel: {
         ...data,
@@ -106,7 +131,7 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
               <InfoField
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.type}
-                value={vessel.type}
+                value={vessel.vesselType}
                 valuesHistory={vessel.history.vesselType.byDate}
               ></InfoField>
               <InfoField
@@ -125,6 +150,7 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.imo}
                 value={vessel.imo}
+                hideTMTDate={true}
                 valuesHistory={vessel.history.imo.byDate}
               ></InfoField>
               <InfoField
@@ -143,18 +169,21 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.length}
                 value={vessel.length}
+                hideTMTDate={true}
                 valuesHistory={vessel.history.length.byDate}
               ></InfoField>
               <InfoField
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.grossTonnage}
                 value={vessel.grossTonnage}
+                hideTMTDate={true}
                 valuesHistory={vessel.history.grossTonnage.byDate}
               ></InfoField>
               <InfoField
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.depth}
                 value={vessel.depth}
+                hideTMTDate={true}
                 valuesHistory={vessel.history.depth.byDate}
               ></InfoField>
               <div className={styles.identifierField}>
@@ -177,6 +206,7 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
               <InfoField
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.builtYear}
+                hideTMTDate={true}
                 value={vessel.builtYear}
               ></InfoField>
               <InfoField
@@ -249,7 +279,7 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
               />
             </Fragment>
           )}
-          {vessel && !offlineVessel && (
+          {vessel && !isMergedVesselsView && !offlineVessel && (
             <Button
               className={styles.saveButton}
               type="secondary"
@@ -259,8 +289,16 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
               {t('vessel.saveForOfflineView', 'SAVE VESSEL FOR OFFLINE VIEW')}
             </Button>
           )}
+          {isMergedVesselsView && (
+            <Button className={styles.saveButton} type="secondary" disabled={true}>
+              {t(
+                'vessel.offlineStillNotAllowedOnMergedVessels',
+                'OFFLINE VIEW NOT ALLOWED ON MERGED VESSELS'
+              )}
+            </Button>
+          )}
         </div>
-        <Highlights></Highlights>
+        <Highlights onMoveToMap={props.onMoveToMap}></Highlights>
       </div>
     </Fragment>
   )