+++ applications/vessel-history/src/features/search/search.thunk.ts
@@ -1,5 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { stringify } from 'qs'
+import { event as uaEvent } from 'react-ga'
 import GFWAPI, {
   AdvancedSearchQueryField,
   getAdvancedSearchQuery,
@@ -7,10 +8,10 @@ import GFWAPI, {
 import { VesselSearch } from '@globalfishingwatch/api-types'
 import { BASE_DATASET, RESULTS_PER_PAGE, SEARCH_MIN_CHARACTERS } from 'data/constants'
 import { RootState } from 'store'
+import { SearchResults } from 'types'
 import { CachedVesselSearch } from './search.slice'
 
 export const getSerializedQuery = (query: string, advancedSearch?: Record<string, any>) => {
-  if (!advancedSearch) return query
   const fields: AdvancedSearchQueryField[] = [
     {
       key: 'shipname',
@@ -18,27 +19,27 @@ export const getSerializedQuery = (query: string, advancedSearch?: Record<string
     },
     {
       key: 'mmsi',
-      value: advancedSearch.mmsi,
+      value: advancedSearch?.mmsi,
     },
     {
       key: 'imo',
-      value: advancedSearch.imo,
+      value: advancedSearch?.imo,
     },
     {
       key: 'callsign',
-      value: advancedSearch.callsign,
+      value: advancedSearch?.callsign,
     },
     {
       key: 'flag',
-      value: advancedSearch.flags.map((f: string) => ({ id: f })),
+      value: advancedSearch?.flags?.map((f: string) => ({ id: f })),
     },
     {
       key: 'lastTransmissionDate',
-      value: advancedSearch.lastTransmissionDate,
+      value: advancedSearch?.lastTransmissionDate,
     },
     {
       key: 'firstTransmissionDate',
-      value: advancedSearch.firstTransmissionDate,
+      value: advancedSearch?.firstTransmissionDate,
     },
   ]
   return getAdvancedSearchQuery(fields)
@@ -60,7 +61,7 @@ const fetchData = async (
     useTMT: true,
   })
 
-  const url = `/v1/vessels/advanced-search?${urlQuery}`
+  const url = `/v1/vessels/advanced-search-tmt?${urlQuery}`
 
   return await GFWAPI.fetch<any>(url, {
     signal,
@@ -110,6 +111,35 @@ export type VesselSearchThunk = {
   advancedSearch?: Record<string, any>
 }
 
+const trackData = (
+  query: any,
+  results: SearchResults | null,
+  actualResults: number
+) => {
+  if (!query.offset || query.offset === 0) {
+    const vessels = results?.vessels.slice(0, 5).map(vessel => {
+      return {
+        gfw: vessel.id,
+        tmt: vessel.vesselMatchId
+      }
+    })
+    uaEvent({
+      category: 'Search Vessel VV',
+      action: 'Click Search',
+      label: JSON.stringify({ ...query, vessels }),
+      value: results?.total
+    })
+  } else {
+    uaEvent({
+      category: 'Search Vessel VV',
+      action: 'Click Load More',
+      label: actualResults.toString(),
+      value: results?.total
+    })
+  }
+
+}
+
 export const fetchVesselSearchThunk = createAsyncThunk(
   'search/vessels',
   async (
@@ -117,6 +147,8 @@ export const fetchVesselSearchThunk = createAsyncThunk(
     { rejectWithValue, getState, signal }
   ) => {
     const searchData = await fetchData(query, offset, signal, advancedSearch)
+    trackData({ query: query, ...advancedSearch }, searchData, 5)
+
     return searchData
   },
   {