+++ applications/vessel-history/src/types/index.ts
@@ -1,4 +1,4 @@
-import { Vessel, Authorization } from '@globalfishingwatch/api-types'
+import { Vessel, Authorization, VesselSearch } from '@globalfishingwatch/api-types'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 
 export type WorkspaceViewportParam = 'latitude' | 'longitude' | 'zoom'
@@ -12,15 +12,21 @@ export type WorkspaceAdvancedSearchParam =
   | 'lastTransmissionDate'
   | 'lastTransmissionDate'
   | 'firstTransmissionDate'
+export type WorkspaceMergeVesselsParam = 'aka'
 
 export type WorkspaceParam =
   | WorkspaceViewportParam
   | WorkspaceTimeRangeParam
   | WorkspaceStateProperty
   | WorkspaceAdvancedSearchParam
+  | WorkspaceMergeVesselsParam
 
 export type WorkspaceViewport = Record<WorkspaceViewportParam, number>
 export type WorkspaceTimeRange = Record<WorkspaceTimeRangeParam, string>
+export type WorkspaceMergeVessels = {
+  aka?: string[]
+}
+
 export type BivariateDataviews = [string, string]
 
 export type WorkspaceState = {
@@ -29,7 +35,10 @@ export type WorkspaceState = {
   dataviewInstances?: Partial<UrlDataviewInstance[]>
   vessel?: string
 }
-export type QueryParams = Partial<WorkspaceViewport> & Partial<WorkspaceTimeRange> & WorkspaceState
+export type QueryParams = Partial<WorkspaceViewport> &
+  Partial<WorkspaceTimeRange> &
+  Partial<WorkspaceMergeVessels> &
+  WorkspaceState
 
 export type CoordinatePosition = {
   latitude: number
@@ -69,6 +78,7 @@ export interface VesselFieldsHistory {
 export interface VesselWithHistory extends Vessel {
   history: VesselFieldsHistory
   iuuStatus?: number
+  vesselType?: string
 }
 
 export enum VesselAPISource {
@@ -158,6 +168,7 @@ export type GFWDetail = {
   dataset: string
   geartype: string
   normalized_shipname: string
+  vesselType: string
 }
 
 export enum Locale {
@@ -183,3 +194,11 @@ export type Range = {
 
 // minX, minY, maxX, maxY
 export type Bbox = [number, number, number, number]
+
+export interface SearchResults {
+  vessels: Array<VesselSearch>
+  query: string
+  offset: number
+  total: number
+  searching: boolean
+}