+++ applications/tile-inspector/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tile-inspector
 
+## 0.1.89
+
+### Patch Changes
+
+- Updated dependencies [06a4320f]
+  - @globalfishingwatch/layer-composer@9.3.0
+
+## 0.1.88
+
+### Patch Changes
+
+- Updated dependencies [52d30f29]
+  - @globalfishingwatch/layer-composer@9.2.3
+
 ## 0.1.87
 
 ### Patch Changes