+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @globalfishingwatchapp/vessel-history
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [06a4320f]
+  - @globalfishingwatch/layer-composer@9.3.0
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [4de42baa]
+- Updated dependencies [52d30f29]
+  - @globalfishingwatch/dataviews-client@13.1.0
+  - @globalfishingwatch/layer-composer@9.2.3
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [30b3034b]
+  - @globalfishingwatch/api-types@5.3.2
+
 ## 1.0.10
 
 ### Patch Changes