+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/layer-composer
 
+## 9.3.0
+
+### Minor Changes
+
+- 06a4320f: support group for heatmap layers
+
+## 9.2.3
+
+### Patch Changes
+
+- 52d30f29: ensure api param {{ is replaced by {
+
 ## 9.2.2
 
 ### Patch Changes